With the proxy decoupled, cleanup some interfaces.
